+++ CHANGELOG.md
@@ -17,6 +17,7 @@ Scott R. Shinn (https://www.atomicorp.com)
 - @bchurchill
 - @alex-front
 - @crlorentzen
+- @hcw2016
 
 **Release Notes**
 
@@ -28,6 +29,7 @@ Work toward the next minor release after 4.2.0.
 - @awiddersheim / @atomicturtle - [PR 578](https://github.com/ossec/ossec-hids/pull/578) - Refactor Unix MQ start/send retry loops without changing wait budgets
 - @bchurchill / @atomicturtle - [PR 633](https://github.com/ossec/ossec-hids/pull/633) - Enable Linux compile/link hardening by default (PIE, FORTIFY, stack protector, full RELRO)
 - @alex-front / @atomicturtle - [PR 1036](https://github.com/ossec/ossec-hids/pull/1036) - Add cPanel/cpsrvd decoders and rules (login, session, access)
+- @hcw2016 / @atomicturtle - [PR 1166](https://github.com/ossec/ossec-hids/pull/1166) - Backup agent state before remove/force-delete and alert on duplicated IP
 
 **Bug Fixes**